% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FindRootCell}
\alias{FindRootCell}
\title{Find Root Cell}
\usage{
FindRootCell(seu = NULL, gene = "Pax6", celltype = "RGC", sample = NULL)
}
\arguments{
\item{seu}{seurat}

\item{gene}{gene}

\item{celltype}{cell type}

\item{sample}{sample name}
}
\description{
Find Root Cell
}
