% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{MakePTMatrix}
\alias{MakePTMatrix}
\title{MakePTMatrix}
\usage{
MakePTMatrix(slingX = "sce_slingshot", counts = NULL, n = 200, lineage = 1)
}
\arguments{
\item{slingX}{sling object}

\item{counts}{count matrix}

\item{n}{total genes}

\item{lineage}{lineage model}
}
\value{
pt-matrix
}
\description{
MakePTMatrix
}
