% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MakePTMatrixSlingshot}
\alias{MakePTMatrixSlingshot}
\title{MakePTMatrix - Slingshot}
\usage{
MakePTMatrixSlingshot(
  slingX = "sce_slingshot",
  counts = NULL,
  nknots = 10,
  n = 200,
  lineage = 1
)
}
\arguments{
\item{slingX}{sling object}

\item{counts}{count matrix}

\item{n}{total genes}

\item{lineage}{lineage model}
}
\value{
pt-matrix
}
\description{
MakePTMatrix - Slingshot
}
