% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{Monocle3MakeCDS}
\alias{Monocle3MakeCDS}
\title{RunMonocle3}
\usage{
Monocle3MakeCDS(obj = NULL, assay = "SCT", umap = "wnn.umap")
}
\arguments{
\item{obj}{seurat object}

\item{assay}{start root}

\item{umap}{out dir}
}
\description{
RunMonocle3
}
