% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{PlotSlingshot}
\alias{PlotSlingshot}
\title{Plot-Slingshot}
\usage{
PlotSlingshot(
  sce_slingshot = NULL,
  lineage = 1,
  w = 4.5,
  h = 3,
  outdir = "out_slingshot"
)
}
\arguments{
\item{sce_slingshot}{sce_slingshot}

\item{lineage}{lineage n}
}
\description{
Plot-Slingshot
}
