% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunM3DropFeatureSelection}
\alias{RunM3DropFeatureSelection}
\title{RunM3DropFeatureSelection}
\usage{
RunM3DropFeatureSelection(
  seu = NULL,
  assay = "SCT",
  is_counts = FALSE,
  outdir = "."
)
}
\arguments{
\item{seu}{seurat}

\item{outdir}{out dir}

\item{root}{start root}
}
\description{
RunM3DropFeatureSelection
}
