% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{RunMonocle2}
\alias{RunMonocle2}
\title{RunMonocle2}
\usage{
RunMonocle2(seu = NULL, root = NULL, outdir = ".")
}
\arguments{
\item{seu}{seurat}

\item{root}{start root}

\item{outdir}{out dir}
}
\description{
RunMonocle2
}
