% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{RunMonocle3Pipe}
\alias{RunMonocle3Pipe}
\title{RunMonocle3Pipe}
\usage{
RunMonocle3Pipe(seu = NULL, root_cells = NULL, umap = "umap", outdir = ".")
}
\arguments{
\item{seu}{seurat object}

\item{root_cells}{start root}

\item{outdir}{out dir}
}
\description{
RunMonocle3Pipe
}
