% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{RunPhate}
\alias{RunPhate}
\title{RunPhate}
\usage{
RunPhate(
  seu = NULL,
  use_py = "/Users/hua/opt/miniconda3/envs/phate/bin/python",
  assay = "SCT",
  seed = 42
)
}
\arguments{
\item{seu}{seurat}

\item{use_py}{python path}

\item{assay}{seurat assay}
}
\value{
seurat object
}
\description{
RunPhate
}
