% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{RunSlingshotPipe_Seurat}
\alias{RunSlingshotPipe_Seurat}
\title{RunSlingshotPipe_Seurat}
\usage{
RunSlingshotPipe_Seurat(
  seu = NULL,
  assay = "SCT",
  reduction = "RNA.UMAP",
  root = "5",
  clus = "seurat_clusters",
  outdir = "out_slingshot",
  w = 4.5,
  h = 3,
  d_marker = "",
  avg_log2FC = 1,
  pct_1 = 0.5,
  diff_pct = 0.3,
  n = 200,
  nknots = 10,
  lineage = 1,
  km = 4
)
}
\arguments{
\item{seu}{pt-matrix}

\item{d_marker}{data frame for markers}
}
\description{
RunSlingshotPipe_Seurat
}
