% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{RunSlingshotSeurat}
\alias{RunSlingshotSeurat}
\title{RunSlingshotSeurat}
\usage{
RunSlingshotSeurat(
  seu = NULL,
  assay = "SCT",
  reduction = "RNA.UMAP",
  root = "5",
  clus = "seurat_clusters",
  w = 4.5,
  h = 3,
  outdir = "out_slingshot"
)
}
\arguments{
\item{seu}{seurat}

\item{assay}{assay}

\item{reduction}{reducedDim}

\item{root}{root name}

\item{clus}{cluster labels}
}
\description{
RunSlingshotSeurat
}
