% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urd.R
\name{RunURD}
\alias{RunURD}
\title{Run URD}
\usage{
RunURD(seu = NULL, cluster = "cell_type2", root = "RGC", seed = 42)
}
\arguments{
\item{seu}{seurat}

\item{cluster}{cluster}

\item{root}{start root}
}
\value{
urd object
}
\description{
Run URD
}
