% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{ScatterPlot2D}
\alias{ScatterPlot2D}
\title{Score_UMAPPlot}
\usage{
ScatterPlot2D(
  data = NULL,
  x = "differentiation_score",
  y = "progenitor_score",
  group = "group",
  x_lab = "Differentiation Score",
  y_lab = "Progenitor Score",
  color = NULL
)
}
\arguments{
\item{data}{data frame}

\item{group}{group}

\item{color}{new color}
}
\value{
plot
}
\description{
Score_UMAPPlot
}
