% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{Score_UMAPPlot}
\alias{Score_UMAPPlot}
\title{Score_UMAPPlot}
\usage{
Score_UMAPPlot(obj = NULL, umap = "umap", name = "")
}
\arguments{
\item{obj}{seurat object}

\item{umap}{umap}

\item{name}{score name}
}
\value{
plot
}
\description{
Score_UMAPPlot
}
