% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{Seurat_EstimateCycling}
\alias{Seurat_EstimateCycling}
\title{Seurat_EstimateCycling}
\usage{
Seurat_EstimateCycling(obj = NULL, ref = "human")
}
\arguments{
\item{obj}{seurat object}

\item{ref}{human/mouse}
}
\value{
data frame
}
\description{
Seurat_EstimateCycling
}
