% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{Seurat_FindMarkers}
\alias{Seurat_FindMarkers}
\title{Seurat_FindMarkers}
\usage{
Seurat_FindMarkers(
  obj = NULL,
  groupby = "cell_type2",
  ident1 = "",
  min_pct = 0.3,
  logfc = 1,
  recorrect_umi = TRUE
)
}
\arguments{
\item{obj}{seurat}

\item{groupby}{cell_type2}

\item{outdir}{out dir}
}
\description{
Seurat_FindMarkers
}
