% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{SeuratAddScore}
\alias{SeuratAddScore}
\title{SeuratAddScore}
\usage{
SeuratAddScore(obj = NULL, name = "", features = "")
}
\arguments{
\item{obj}{seurat object}

\item{name}{score name}

\item{features}{gene set}
}
\value{
seurat obj
}
\description{
SeuratAddScore
}
