% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{RunComplexHeatmap}
\alias{RunComplexHeatmap}
\title{RunComplexHeatmap}
\usage{
RunComplexHeatmap(mtx = "pt.matrix", km = 4, outdir = ".")
}
\arguments{
\item{mtx}{pt-matrix}

\item{km}{km}
}
\value{
ht
}
\description{
RunComplexHeatmap
}
