% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{RunSlingshotSeurat}
\alias{RunSlingshotSeurat}
\title{RunSlingshotSeurat}
\usage{
RunSlingshotSeurat(
  seu = NULL,
  assay = "SCT",
  ident = "seurat_clusters",
  reduction = "PCA",
  start_root = NULL,
  allow_breaks = TRUE,
  save_out = FALSE,
  outdir = "out_slingshot"
)
}
\arguments{
\item{seu}{seurat}

\item{assay}{assay}

\item{ident}{ident labels}

\item{reduction}{reducedDim}

\item{start_root}{root name}
}
\description{
RunSlingshotSeurat
}
