% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleLineage.R
\name{RunTriangleAnalysis}
\alias{RunTriangleAnalysis}
\title{RunTriangleAnalysis}
\usage{
RunTriangleAnalysis(
  seu = NULL,
  ref = "human",
  out_dir = "out_triangelplot",
  root = "CycProg",
  armL = "Astrocyte",
  armR = "Neuron",
  min_pct_shared = 0.8,
  logfc_shared = 0.05,
  min_pct_armL = 0.5,
  logfc_armL = 1,
  min_pct_armR = 0.5,
  logfc_armR = 0.5,
  top_n_genes = 50,
  expr_ratio = 0.8
)
}
\arguments{
\item{seu}{Seurat object}

\item{ref}{human/mouse}

\item{root}{root cell type}

\item{armL}{left arm cell type}

\item{armR}{right arm cell type}

\item{min_pct_shared}{pct cutoff for shared}

\item{logfc_shared}{logfc cutoff for shared}

\item{min_pct_armL}{pct cutoff for armL}

\item{logfc_armL}{logfc cutoff for armL}

\item{min_pct_armR}{pct cutoff for armR}

\item{logfc_armR}{logfc cutoff for armR}

\item{top_n_genes}{cutoff top genes}

\item{expr_ratio}{cutoff ratio of expressed cells}

\item{outdir}{out dir.}
}
\value{
data frame
}
\description{
RunTriangleAnalysis
}
